V2 - Source code from 2012 archive
Finally found V2!

This code is here for archive purposes. I have actually fixed this to
run just so that I could verify it was indeed v2, so as long as the
powerpack nuget package is installed then this will be possible to run.

This is a very buggy program so don't expect anything to really work,
especially adding bookmarks (which tend to freeze up, and a bug that's
present that started in V1). I also had to edit this slightly since a
timer handler had a conflicting for loop variable name. Otherwise this
is as original as it gets.